Fix possible overflow
CWE-190, CWE-192, CWE-197, CWE-681

+++ src/core/StelHips.cpp
@@ -309,7 +309,7 @@ HipsTile* HipsSurvey::getTile(int order, int pix)
 			tile->allsky = texMgr.createTexture(image, StelTexture::StelTextureParams(true));
 		}
 		int tileWidth = getPropertyInt("hips_tile_width", 512);
-		tiles.insert(uid, tile, tileWidth * tileWidth);
+		tiles.insert(uid, tile, static_cast<long>(tileWidth) * tileWidth);
 	}
 	return tile;
 }
